[premake] Reorganized projects, added another doc file 	Using premake-dev to generate make files/ solutions

+++ src/Core/LLAPI/Parallel/Thread.h
@@ -20,7 +20,7 @@ namespace Oak3D
 			~Thread();
 			
 			void Join();
-			void Sleep(uint64_t ms);
+			void Sleep(uint32_t ms);
 		private:
 			struct ThreadFunc
 			{
